# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FLOZ Platform API",
        "description": "API documentation for FLOZ Platform and Tenant specific features",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/tenant/grades": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "List Grades",
                "description": "Get list of grades with filtering",
                "operationId": "038f0c57166ca8a5f1d689df03a31a04",
                "parameters": [
                    {
                        "name": "class_id",
                        "in": "query",
                        "description": "Filter by class ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "semester_id",
                        "in": "query",
                        "description": "Filter by semester ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "subject_id",
                        "in": "query",
                        "description": "Filter by subject ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of grades"
                    }
                }
            }
        },
        "/tenant/grades/batch": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Batch Input View",
                "description": "Get view for batch grade input",
                "operationId": "7eb2dc875cba15929eac37b954a35299",
                "parameters": [
                    {
                        "name": "class_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "semester_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "subject_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Batch input view"
                    }
                }
            },
            "post": {
                "tags": [
                    "Grades"
                ],
                "summary": "Store Batch Grades",
                "description": "Store multiple grades at once",
                "operationId": "0443a33d2c21f496e078e2e21dead376",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "class_id",
                                    "semester_id",
                                    "subject_id",
                                    "grades"
                                ],
                                "properties": {
                                    "class_id": {
                                        "type": "integer"
                                    },
                                    "semester_id": {
                                        "type": "integer"
                                    },
                                    "subject_id": {
                                        "type": "integer"
                                    },
                                    "grades": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "student_id": {
                                                    "type": "integer"
                                                },
                                                "daily_test_avg": {
                                                    "type": "number"
                                                },
                                                "mid_test": {
                                                    "type": "number"
                                                },
                                                "final_test": {
                                                    "type": "number"
                                                },
                                                "knowledge_score": {
                                                    "type": "number"
                                                },
                                                "skill_score": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect to index"
                    }
                }
            }
        },
...
    },
...
}